Change nSpendHeight and GetSpendHeight to int64_t (fixes H2)
int nSpendHeight in CheckTxInputs caused signed/unsigned mismatch with
coin.nHeight (uint32_t:31). C++ arithmetic conversions promoted int to
unsigned, so a negative or underflowing nSpendHeight could produce a
huge positive depth, silently bypassing the coinbase/pegout maturity
checks. Changing to int64_t ensures coin.nHeight is promoted upward to
int64_t instead, making underflow impossible and negative values
detectable. GetSpendHeight updated to match; fuzz caller updated to use
ConsumeIntegralInRange<int64_t>. Consistent with the H1 fix that moved
block-height fields in consensus/params.h to int64_t.

+++ src/consensus/tx_verify.cpp
@@ -166,7 +166,7 @@ int64_t GetTransactionSigOpCost(const CTransaction& tx, const CCoinsViewCache& i
     return nSigOps;
 }
 
-bool Consensus::CheckTxInputs(const CTransaction& tx, TxValidationState& state, const CCoinsViewCache& inputs, int nSpendHeight, CAmount& txfee)
+bool Consensus::CheckTxInputs(const CTransaction& tx, TxValidationState& state, const CCoinsViewCache& inputs, int64_t nSpendHeight, CAmount& txfee)
 {
     // are the actual inputs available?
     if (!inputs.HaveInputs(tx)) {
